// src/lib/shapes/types.ts
// Kept free of React so lib stays renderer agnostic.
export enum ShapeTypes {
  Line = "Line",
  CurveLine = "CurveLine",
  Circle = "Circle",
  Rectangle = "Rectangle",
  RoundedRectangle = "Rounded Rectangle",
  Polygon = "Polygon",
  Triangle = "Triangle",
  RightTriangle = "Right Triangle",
  Diamond = "Diamond",
  Pentagon = "Pentagon",
  Hexagon = "Hexagon",
  RightArrow = "Right Arrow",
  LeftArrow = "Left Arrow",
  UpArrow = "Up Arrow",
  DownArrow = "Down Arrow",
  FourPointStar = "Four-Point Star",
  FivePointStar = "Five-Point Star",
  SixPointStar = "Six-Point Star",
  RoundedRectangularCallout = "Rounded Rectangular Callout",
  OvalCallout = "Oval Callout",
  Heart = "Heart",
}